+++ gdal/ogr/ogrsf_frmts/gml/gmlreader.cpp
@@ -791,20 +791,20 @@ int GMLReader::GetFeatureElementIndex( const char *pszElement, int nElementLengt
         }
 
         /* Begin of CSW SearchResults */
-        else if (nElementLength == strlen("BriefRecord") &&
-                 nLenLast == strlen("SearchResults") &&
+        else if (nElementLength == (int)strlen("BriefRecord") &&
+                 nLenLast == (int)strlen("SearchResults") &&
                  strcmp(pszElement, "BriefRecord") == 0 &&
                  strcmp(pszLast, "SearchResults") == 0)
         {
         }
-        else if (nElementLength == strlen("SummaryRecord") &&
-                 nLenLast == strlen("SearchResults") &&
+        else if (nElementLength == (int)strlen("SummaryRecord") &&
+                 nLenLast == (int)strlen("SearchResults") &&
                  strcmp(pszElement, "SummaryRecord") == 0 &&
                  strcmp(pszLast, "SearchResults") == 0)
         {
         }
-        else if (nElementLength == strlen("Record") &&
-                 nLenLast == strlen("SearchResults") &&
+        else if (nElementLength == (int)strlen("Record") &&
+                 nLenLast == (int)strlen("SearchResults") &&
                  strcmp(pszElement, "Record") == 0 &&
                  strcmp(pszLast, "SearchResults") == 0)
         {
